Push "Achievement" model and UV
Model contains FOUR achievements. each take their respective slot on the
UV Maping.
achievement1 through achievement4 can be called individually. Depending
on the texture wanted, an RGB may need to be created and assigned in the
egg file. If not, then the BAM should be good to go as is.