Tunnel work
added the type <no_touch> to the LinkTunnel1 node and to the
tunnel_sphere node. If the tunnel isn't working do to flattening, I
would think this setting set should eliminate those to nodes from being
messed with. If this don't work now in Live, it's got to be something in
the code. I tested it locally and literally entered/exited no less than
30 times flawlessly.